+++ mod-record-specifications-server/src/main/java/org/folio/rspec/integration/kafka/SpecificationFullChangedEventProducer.java
@@ -1,21 +1,22 @@
 package org.folio.rspec.integration.kafka;
 
-import java.util.UUID;
 import org.folio.rspec.domain.dto.SpecificationUpdatedEvent;
+import org.folio.rspec.domain.repository.SpecificationRepository;
 import org.folio.spring.FolioExecutionContext;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Component;
 
 @Component("fullChangeProducer")
 public class SpecificationFullChangedEventProducer extends SpecificationPartialChangedEventProducer {
-  public SpecificationFullChangedEventProducer(
-    KafkaTemplate<String, SpecificationUpdatedEvent> template,
-    FolioExecutionContext context) {
-    super(template, context);
+
+  public SpecificationFullChangedEventProducer(KafkaTemplate<String, SpecificationUpdatedEvent> template,
+                                               SpecificationRepository specificationRepository,
+                                               FolioExecutionContext context) {
+    super(template, specificationRepository, context);
   }
 
   @Override
-  protected SpecificationUpdatedEvent buildEvent(UUID specificationId) {
-    return new SpecificationUpdatedEvent(specificationId, tenantId(), SpecificationUpdatedEvent.UpdateExtent.FULL);
+  protected SpecificationUpdatedEvent.UpdateExtent updateExtent() {
+    return SpecificationUpdatedEvent.UpdateExtent.FULL;
   }
 }